MC-38185: [Magento Cloud] Issue with the new In store pickup when we have multiple languages.
